Refactor: Channel 구현을 위한 Context 분리 등
- useContext.tsx 파일을 useUserContext, useAppContext로 분리
  - useContext.tsx 파일은 ContextProvider 용도로만 사용
  - AppStateType에 socket 추가
- 로그인 완료 시 socket 연결 로직 추가
- ListItem.stories.tsx 파일명 변경
