+++ package.json
@@ -20,6 +20,7 @@
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "react-toastify": "^7.0.4",
+    "socket.io-client": "^4.2.0",
     "typescript": "^4.1.2",
     "web-vitals": "^1.0.1"
   },
