+++ src/components/atoms/ListItem/ListItem.stories.tsx
@@ -8,7 +8,7 @@ import Typo from '../Typo/Typo';
 import Button from '../Button/Button';
 import List from '../List/List';
 import MainTemplate from '../../templates/MainTemplate/MainTemplate';
-import { ContextProvider } from '../../../utils/hooks/useContext';
+import ContextProvider from '../../../utils/hooks/useContext';
 import ProfileCard from '../../organisms/ProfileCard/ProfileCard';
 import { RelatedInfoType } from '../../../types/User';
