+++ src/components/molecules/Menu/Menu.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import { Meta } from '@storybook/react';
 import Menu from './Menu';
-import { ContextProvider } from '../../../utils/hooks/useContext';
+import ContextProvider from '../../../utils/hooks/useContext';
 
 export default {
   title: 'molecules/Menu',
