+++ src/components/organisms/ProfileCard/ProfileCard.tsx
@@ -4,7 +4,8 @@ import axios from 'axios';
 import { toast } from 'react-toastify';
 import Grid from '@material-ui/core/Grid';
 import { makeStyles } from '@material-ui/core/styles';
-import { useAppDispatch, useUserState } from '../../../utils/hooks/useContext';
+import { useUserState } from '../../../utils/hooks/useUserContext';
+import { useAppDispatch } from '../../../utils/hooks/useAppContext';
 import { FriendshipType, RelatedInfoType } from '../../../types/User';
 import Button from '../../atoms/Button/Button';
 import UserProfile from '../../molecules/UserProfile/UserProfile';
