+++ src/components/organisms/UserInfoForm/UserInfoForm.tsx
@@ -7,7 +7,8 @@ import { useHistory } from 'react-router-dom';
 import Typo from '../../atoms/Typo/Typo';
 import Avatar from '../../atoms/Avatar/Avatar';
 import { makeAPIPath } from '../../../utils/utils';
-import { useAppDispatch, useUserDispatch } from '../../../utils/hooks/useContext';
+import { useUserDispatch } from '../../../utils/hooks/useUserContext';
+import { useAppDispatch } from '../../../utils/hooks/useAppContext';
 import Input from '../../atoms/Input/Input';
 import Button from '../../atoms/Button/Button';
 import Switch from '../../atoms/Switch/Switch';
