+++ src/components/pages/LoginPage/LoginPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useAppDispatch } from '../../../utils/hooks/useContext';
+import { useAppDispatch } from '../../../utils/hooks/useAppContext';
 import { makeAPIPath } from '../../../utils/utils';
 import Button from '../../atoms/Button/Button';
 import Typo from '../../atoms/Typo/Typo';
