+++ src/components/pages/MFAPage/MFAPage.stories.tsx
@@ -3,7 +3,7 @@ import { Meta } from '@storybook/react';
 import { ToastContainer } from 'react-toastify';
 import { BrowserRouter } from 'react-router-dom';
 import MFAPage from './MFAPage';
-import { ContextProvider } from '../../../utils/hooks/useContext';
+import ContextProvider from '../../../utils/hooks/useContext';
 
 export default {
   component: MFAPage,
