+++ src/components/pages/MFAPage/MFAPage.tsx
@@ -4,7 +4,8 @@ import { makeStyles } from '@material-ui/core/styles';
 import Grid from '@material-ui/core/Grid';
 import { useHistory } from 'react-router-dom';
 import { toast } from 'react-toastify';
-import { useAppDispatch, useUserDispatch, useUserState } from '../../../utils/hooks/useContext';
+import { useUserDispatch, useUserState } from '../../../utils/hooks/useUserContext';
+import { useAppDispatch } from '../../../utils/hooks/useAppContext';
 import { makeAPIPath } from '../../../utils/utils';
 import Button from '../../atoms/Button/Button';
 import DigitInput from '../../atoms/DigitInput/DigitInput';
