+++ src/components/pages/ProfilePage/ProfilePage.tsx
@@ -3,7 +3,8 @@ import { RouteComponentProps, useHistory } from 'react-router-dom';
 import { toast } from 'react-toastify';
 import Grid from '@material-ui/core/Grid';
 import { RelatedInfoType } from '../../../types/User';
-import { useAppDispatch, useUserState } from '../../../utils/hooks/useContext';
+import { useUserState } from '../../../utils/hooks/useUserContext';
+import { useAppDispatch } from '../../../utils/hooks/useAppContext';
 import { asyncGetRequest, makeAPIPath } from '../../../utils/utils';
 import List from '../../atoms/List/List';
 import Typo from '../../atoms/Typo/Typo';
