+++ src/components/pages/RegisterPage/RegisterPage.stories.tsx
@@ -4,7 +4,7 @@ import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import { Meta } from '@storybook/react';
 import RegisterPage from './RegisterPage';
-import { ContextProvider } from '../../../utils/hooks/useContext';
+import ContextProvider from '../../../utils/hooks/useContext';
 
 export default {
   component: RegisterPage,
