+++ src/components/templates/MainTemplate/MainTemplate.stories.tsx
@@ -3,7 +3,7 @@ import { Meta } from '@storybook/react';
 import { BrowserRouter } from 'react-router-dom';
 import { makeStyles } from '@material-ui/core/styles';
 import MainTemplate from './MainTemplate';
-import { ContextProvider } from '../../../utils/hooks/useContext';
+import ContextProvider from '../../../utils/hooks/useContext';
 
 export default {
   component: MainTemplate,
